+++ src/main/kotlin/no/nav/familie/pdf/pdf/domain/PdfMedStandarder.kt
@@ -1,0 +1,6 @@
+package no.nav.familie.pdf.pdf.domain
+
+data class PdfMedStandarder(
+    val pdf: ByteArray,
+    val standarder: Standarder,
+)
